+++ Source/Client/Sprites.h
@@ -105,8 +105,8 @@ struct MapSprite
     uint Color {};
     uint ContourColor {};
     bool IsTweakOffs {};
-    short TweakOffsX {};
-    short TweakOffsY {};
+    int TweakOffsX {};
+    int TweakOffsY {};
     bool IsTweakAlpha {};
     uchar TweakAlpha {};
 };
@@ -145,8 +145,8 @@ public:
     int ScrY {};
     int* PScrX {};
     int* PScrY {};
-    short* OffsX {};
-    short* OffsY {};
+    int* OffsX {};
+    int* OffsY {};
     Sprite* Parent {};
     Sprite* Child {};
     uchar* Alpha {};
@@ -186,15 +186,15 @@ public:
     [[nodiscard]] auto RootSprite() -> Sprite*;
     [[nodiscard]] auto Size() const -> uint;
 
-    [[nodiscard]] auto AddSprite(DrawOrderType draw_order, ushort hx, ushort hy, int x, int y, int* sx, int* sy, uint id, uint* id_ptr, short* ox, short* oy, uchar* alpha, RenderEffect** effect, bool* callback) -> Sprite&;
-    [[nodiscard]] auto InsertSprite(DrawOrderType draw_order, ushort hx, ushort hy, int x, int y, int* sx, int* sy, uint id, uint* id_ptr, short* ox, short* oy, uchar* alpha, RenderEffect** effect, bool* callback) -> Sprite&;
+    [[nodiscard]] auto AddSprite(DrawOrderType draw_order, ushort hx, ushort hy, int x, int y, int* sx, int* sy, uint id, uint* id_ptr, int* ox, int* oy, uchar* alpha, RenderEffect** effect, bool* callback) -> Sprite&;
+    [[nodiscard]] auto InsertSprite(DrawOrderType draw_order, ushort hx, ushort hy, int x, int y, int* sx, int* sy, uint id, uint* id_ptr, int* ox, int* oy, uchar* alpha, RenderEffect** effect, bool* callback) -> Sprite&;
 
     void Unvalidate();
     void SortByMapPos();
     void Clear();
 
 private:
-    [[nodiscard]] auto PutSprite(Sprite* child, DrawOrderType draw_order, ushort hx, ushort hy, int x, int y, int* sx, int* sy, uint id, uint* id_ptr, short* ox, short* oy, uchar* alpha, RenderEffect** effect, bool* callback) -> Sprite&;
+    [[nodiscard]] auto PutSprite(Sprite* child, DrawOrderType draw_order, ushort hx, ushort hy, int x, int y, int* sx, int* sy, uint id, uint* id_ptr, int* ox, int* oy, uchar* alpha, RenderEffect** effect, bool* callback) -> Sprite&;
 
     void GrowPool();
